+++ src/main/java/com/buuz135/industrial/item/addon/CustomAddon.java
@@ -4,9 +4,11 @@ import com.buuz135.industrial.IndustrialForegoing;
 import com.buuz135.industrial.utils.Reference;
 import net.ndrei.teslacorelib.items.BaseAddon;
 
-public class CustomAddon extends BaseAddon {
+public abstract class CustomAddon extends BaseAddon {
 
     protected CustomAddon(String registryName) {
         super(Reference.MOD_ID, IndustrialForegoing.creativeTab, registryName);
     }
+
+    public abstract void createRecipe();
 }