+++ src/main/java/com/buuz135/industrial/proxy/BlockRegistry.java
@@ -61,6 +61,7 @@ public class BlockRegistry {
     public static BlackHoleTankBlock blackHoleTankBlock = new BlackHoleTankBlock();
     public static ResourcefulFurnaceBlock resourcefulFurnaceBlock = new ResourcefulFurnaceBlock();
     public static VillagerTradeExchangerBlock villagerTradeExchangerBlock = new VillagerTradeExchangerBlock();
+    public static EnergyFieldProviderBlock energyFieldProviderBlock = new EnergyFieldProviderBlock();
 
     public static IFCustomFluidBlock BLOCK_ESSENCE = new IFCustomFluidBlock(FluidsRegistry.ESSENCE, Material.WATER);
     public static IFCustomFluidBlock BLOCK_MILK = new IFCustomFluidBlock(FluidsRegistry.MILK, Material.WATER);