+++ src/main/java/com/buuz135/industrial/proxy/ItemRegistry.java
@@ -2,6 +2,7 @@ package com.buuz135.industrial.proxy;
 
 import com.buuz135.industrial.item.*;
 import com.buuz135.industrial.item.addon.AdultFilterAddonItem;
+import com.buuz135.industrial.item.addon.EnergyFieldAddon;
 import com.buuz135.industrial.item.addon.RangeAddonItem;
 import com.buuz135.industrial.utils.RecipeUtils;
 import net.minecraft.item.Item;
@@ -27,7 +28,7 @@ public class ItemRegistry {
 
     public static AdultFilterAddonItem adultFilterAddomItem;
     public static RangeAddonItem rangeAddonItem;
-
+    public static EnergyFieldAddon energyFieldAddon;
 
     public static void registerItems(IForgeRegistry<Item> itemRegistry) {
         (tinyDryRubber = new IFCustomItem("tinydryrubber")).register(itemRegistry);
@@ -46,6 +47,7 @@ public class ItemRegistry {
 
         (adultFilterAddomItem = new AdultFilterAddonItem()).registerItem(itemRegistry);
         (rangeAddonItem = new RangeAddonItem()).registerItem(itemRegistry);
+        (energyFieldAddon = new EnergyFieldAddon()).registerItem(itemRegistry);
 
         meatFeederItem.createRecipe();
         mobImprisonmentToolItem.createRecipe();
@@ -53,5 +55,6 @@ public class ItemRegistry {
         laserLensItem_inverted.createRecipe();
         adultFilterAddomItem.createRecipe();
         rangeAddonItem.createRecipe();
+        energyFieldAddon.createRecipe();
     }
 }