+++ src/main/java/com/buuz135/industrial/proxy/client/ItemRenderRegistry.java
@@ -15,6 +15,7 @@ public class ItemRenderRegistry {
         laserLensItem.registerRender();
         laserLensItem_inverted.registerRender();
         pinkSlime.registerRender();
+        energyFieldAddon.registerRenderer();
 
         adultFilterAddomItem.registerRenderer();
         rangeAddonItem.registerRenderer();