+++ src/main/java/com/buuz135/industrial/registry/IFRegistries.java
@@ -14,5 +14,6 @@ public class IFRegistries {
             .setType(StrawHandler.class)
             .create();
 
-    public static void poke() {}
+    public static void poke() {
+    }
 }